feature(routing): implement go_router navigation migration
